Sentence added

+++ addition.java
@@ -4,7 +4,7 @@ public class addition {
         Scanner userValue = new Scanner(System.in);
     int a = userValue.nextInt();
     int b = userValue.nextInt();
-    int sum = (int) a + b;
+    int sum =  a + b;
 
     System.out.println("The sum of " + a + " and " + b + " is: " +  sum);
     }
